services/agent_speech/app/requirements.txt
fastapi
uvicorn
python-multipart
groq>=1.0.0
python-dotenv